refactor: annotation 기반으로 변경
