+++ slipp/src/main/java/slipp/ManualHandlerMapping.java
@@ -3,10 +3,8 @@ package slipp;
 import nextstep.mvc.DispatcherServlet;
 import nextstep.mvc.HandlerMapping;
 import nextstep.mvc.asis.Controller;
-import nextstep.mvc.asis.ForwardController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import slipp.controller.*;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.HashMap;
@@ -19,15 +17,15 @@ public class ManualHandlerMapping implements HandlerMapping {
     @Override
     public void initialize() {
 //        mappings.put("/", new HomeController());
-        mappings.put("/users/form", new ForwardController("/user/form.jsp"));
-        mappings.put("/users/loginForm", new ForwardController("/user/login.jsp"));
+//        mappings.put("/users/form", new ForwardController(""));
+//        mappings.put("/users/loginForm", new ForwardController("/user/login.jsp"));
 //        mappings.put("/users", new ListUserController());
 //        mappings.put("/users/login", new LoginController());
 //        mappings.put("/users/profile", new ProfileController());
 //        mappings.put("/users/logout", new LogoutController());
 //        mappings.put("/users/create", new CreateUserController());
-        mappings.put("/users/updateForm", new UpdateFormUserController());
-        mappings.put("/users/update", new UpdateUserController());
+//        mappings.put("/users/updateForm", new UpdateFormUserController());
+//        mappings.put("/users/update", new UpdateUserController());
 
         logger.info("Initialized Request Mapping!");
         mappings.keySet().forEach(path -> {
