+++ slipp/src/main/java/slipp/controller/ForwardController.java
@@ -1,0 +1,24 @@
+package slipp.controller;
+
+import nextstep.mvc.tobe.JspView;
+import nextstep.mvc.tobe.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.annotation.RequestMethod;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@Controller
+public class ForwardController {
+
+    @RequestMapping(value = "/users/loginForm", method = RequestMethod.GET)
+    public ModelAndView loginForm(HttpServletRequest request, HttpServletResponse response) {
+        return new ModelAndView(new JspView("/user/login.jsp"));
+    }
+
+    @RequestMapping(value = "/users/form", method = RequestMethod.GET)
+    public ModelAndView userForm(HttpServletRequest request, HttpServletResponse response) {
+        return new ModelAndView(new JspView("/user/form.jsp"));
+    }
+}
