+++ slipp/src/main/java/slipp/controller/UpdateFormUserController.java
@@ -1,6 +1,10 @@
 package slipp.controller;
 
-import nextstep.mvc.asis.Controller;
+import nextstep.mvc.tobe.JspView;
+import nextstep.mvc.tobe.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.annotation.RequestMethod;
 import slipp.controller.exception.NotFoundUserException;
 import slipp.domain.User;
 import slipp.support.db.DataBase;
@@ -8,10 +12,11 @@ import slipp.support.db.DataBase;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class UpdateFormUserController implements Controller {
+@Controller
+public class UpdateFormUserController {
 
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
+    @RequestMapping(value = "/users/updateForm", method = RequestMethod.PUT)
+    public ModelAndView execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
         String userId = req.getParameter("userId");
         User user = DataBase.findUserById(userId)
                 .orElseThrow(NotFoundUserException::new);
@@ -19,7 +24,10 @@ public class UpdateFormUserController implements Controller {
         if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
             throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
         }
-        req.setAttribute("user", user);
-        return "/user/updateForm.jsp";
+
+        ModelAndView modelAndView = new ModelAndView(new JspView("/user/updateForm.jsp"));
+        modelAndView.addObject("user", user);
+
+        return modelAndView;
     }
 }
