+++ slipp/src/main/java/slipp/controller/UpdateUserController.java
@@ -1,20 +1,25 @@
 package slipp.controller;
 
+import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.RedirectView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.annotation.RequestMethod;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import slipp.controller.exception.NotFoundUserException;
 import slipp.domain.User;
 import slipp.support.db.DataBase;
-import nextstep.mvc.asis.Controller;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class UpdateUserController implements Controller {
+@Controller
+public class UpdateUserController {
     private static final Logger log = LoggerFactory.getLogger(UpdateUserController.class);
 
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
+    @RequestMapping(value = "/users/update", method = RequestMethod.PUT)
+    public ModelAndView execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
         User user = DataBase.findUserById(req.getParameter("userId"))
                 .orElseThrow(NotFoundUserException::new);
         if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
@@ -23,8 +28,8 @@ public class UpdateUserController implements Controller {
 
         User updateUser = new User(req.getParameter("userId"), req.getParameter("password"), req.getParameter("name"),
                 req.getParameter("email"));
-        log.debug("Update User : {}", updateUser);
+
         user.update(updateUser);
-        return "redirect:/";
+        return new ModelAndView(new RedirectView("/"));
     }
 }
